Seperate between core MethodChannel and sribuu MethodChannel
