+++ core/android/src/main/kotlin/com/moengage/flutter/sribuu/Constants.kt
@@ -1,0 +1,4 @@
+package com.moengage.flutter.sribuu
+
+const val SRIBUU_MODULE_TAG = "MoEFlutter_sribuu_"
+const val FLUTTER_PLUGIN_SRIBUU_CHANNEL_NAME = "com.moengage/sribuu"
